requirements.txt
discord.py
httpx
google-genai >= 0.7.0 # Ensure a recent version for potential API changes
openai >= 1.3.0 # Ensure a recent version
pyyaml
youtube-transcript-api
google-api-python-client
asyncpraw
beautifulsoup4
google-search-results # SerpAPI library
playwright # Added for custom Google Lens fallback
lxml